Add role creation to the Code Jam creation flow
- Modify the Code Jam creation flow to create roles, and make use of them when setting channel permissions.
